ditch stdint.h where it makes sense

+++ fileio.c
@@ -18,14 +18,14 @@ int fs_seek(FILE *stream, long offset, int whence) {
   return (int) SetFilePointer(stream, offset, 0, (unsigned) whence);
 }
 
-off_t fs_read(void *ptr, size_t size, size_t j, FILE *stream) {
+size_t fs_read(void *ptr, size_t size, size_t j, FILE *stream) {
   DWORD q;
-  return ReadFile(stream, ptr, (unsigned) (size * j), &q, 0);
+  return (size_t) ReadFile(stream, ptr, (unsigned) (size * j), &q, 0);
 }
 
-off_t fs_write(const void *ptr, size_t size, size_t j, FILE *stream) {
+size_t fs_write(const void *ptr, size_t size, size_t j, FILE *stream) {
   DWORD q;
-  return WriteFile(stream, ptr, (unsigned) (size * j), &q, 0);
+  return (size_t) WriteFile(stream, ptr, (unsigned) (size * j), &q, 0);
 }
 
 int fs_close(FILE *file) {
